backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
reportlab>=4.0.0
PyPDF2>=3.0.0
